refactor: centralize exceptions
